by yanniboi: Added DisplayInterface and moved DisplayVariantInterface to Entity namespace.
