+++ modules/mini_panels/src/Form/AccessConditionDeleteForm.php
@@ -8,8 +8,8 @@
 namespace Drupal\mini_panels\Form;
 
 use Drupal\Core\Form\FormStateInterface;
-use Drupal\mini_panels\MiniPanelInterface;
 use Drupal\Core\Form\ConfirmFormBase;
+use Drupal\ctools\Entity\DisplayInterface;
 
 /**
  * Provides a form for deleting an access condition.
@@ -19,7 +19,7 @@ class AccessConditionDeleteForm extends ConfirmFormBase {
   /**
    * The mini_panel entity this selection condition belongs to.
    *
-   * @var \Drupal\mini_panels\MiniPanelInterface
+   * @var \Drupal\ctools\Entity\DisplayInterface
    */
   protected $mini_panel;
 
@@ -61,7 +61,7 @@ class AccessConditionDeleteForm extends ConfirmFormBase {
   /**
    * {@inheritdoc}
    */
-  public function buildForm(array $form, FormStateInterface $form_state, MiniPanelInterface $mini_panel = NULL, $condition_id = NULL) {
+  public function buildForm(array $form, FormStateInterface $form_state, DisplayInterface $mini_panel = NULL, $condition_id = NULL) {
     $this->mini_panel = $mini_panel;
     $this->accessCondition = $mini_panel->getAccessCondition($condition_id);
     return parent::buildForm($form, $form_state);
