+++ src/Form/DisplayVariantDeleteForm.php
@@ -19,7 +19,7 @@ class DisplayVariantDeleteForm extends EntityConfirmFormBase {
   /**
    * The entity being used by this form.
    *
-   * @var \Drupal\ctools\DisplayVariantInterface
+   * @var \Drupal\ctools\Entity\DisplayVariantInterface
    */
   protected $entity;
